backend/ml/requirements.txt
tensorflow==2.13.0
scikit-learn==1.3.0
pandas==2.0.3
numpy==1.24.3
nltk==3.8.1
pytesseract==0.3.10
Pillow==10.0.0
joblib==1.3.2